rename a package from quotecc to the quotecf as well
