+++ src/main/java/com/selina/lending/internal/mapper/quotecf/QuickQuoteCFRequestMapper.java
@@ -1,11 +1,11 @@
-package com.selina.lending.internal.mapper.quotecc;
+package com.selina.lending.internal.mapper.quotecf;
 
 
 import com.selina.lending.internal.dto.quote.QuickQuoteApplicationRequest;
 import com.selina.lending.internal.mapper.LoanInformationMapper;
 import com.selina.lending.internal.mapper.QuickQuoteApplicantMapper;
 import com.selina.lending.internal.mapper.QuickQuotePropertyDetailsMapper;
-import com.selina.lending.internal.service.application.domain.quotecc.QuickQuoteCFRequest;
+import com.selina.lending.internal.service.application.domain.quotecf.QuickQuoteCFRequest;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
