+++ src/main/java/com/selina/lending/internal/mapper/quotecf/QuickQuoteCFResponseMapper.java
@@ -1,7 +1,7 @@
-package com.selina.lending.internal.mapper.quotecc;
+package com.selina.lending.internal.mapper.quotecf;
 
 import com.selina.lending.internal.dto.quote.QuickQuoteResponse;
-import com.selina.lending.internal.service.application.domain.quotecc.QuickQuoteCFResponse;
+import com.selina.lending.internal.service.application.domain.quotecf.QuickQuoteCFResponse;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
