+++ src/main/java/com/selina/lending/internal/service/CreateApplicationService.java
@@ -19,8 +19,8 @@ package com.selina.lending.internal.service;
 
 import com.selina.lending.internal.service.application.domain.ApplicationRequest;
 import com.selina.lending.internal.service.application.domain.ApplicationResponse;
-import com.selina.lending.internal.service.application.domain.quotecc.QuickQuoteCFRequest;
-import com.selina.lending.internal.service.application.domain.quotecc.QuickQuoteCFResponse;
+import com.selina.lending.internal.service.application.domain.quotecf.QuickQuoteCFRequest;
+import com.selina.lending.internal.service.application.domain.quotecf.QuickQuoteCFResponse;
 
 public interface CreateApplicationService {
     ApplicationResponse createDipCCApplication(ApplicationRequest applicationRequest);
