+++ src/main/java/com/selina/lending/internal/service/application/domain/quotecf/QuickQuoteCFResponse.java
@@ -1,4 +1,4 @@
-package com.selina.lending.internal.service.application.domain.quotecc;
+package com.selina.lending.internal.service.application.domain.quotecf;
 
 import com.selina.lending.internal.service.application.domain.Offer;
 import lombok.Builder;
