+++ src/test/java/com/selina/lending/internal/mapper/quotecf/QuickQuoteCFRequestMapperTest.java
@@ -1,7 +1,7 @@
-package com.selina.lending.internal.mapper.quotecc;
+package com.selina.lending.internal.mapper.quotecf;
 
 import com.selina.lending.internal.mapper.MapperBase;
-import com.selina.lending.internal.service.application.domain.quotecc.QuickQuoteCFRequest;
+import com.selina.lending.internal.service.application.domain.quotecf.QuickQuoteCFRequest;
 import org.junit.jupiter.api.Test;
